move some data over and nonmatch GetUnitSum
